refactor: 폴더 구조를 FSD로 변경
- src
    - pages
    - features
    - entities
        - 도메인1
            - components
        - 도메인2
            - components
    - shared
        - components
        - styles
        - types
