+++ frontend/src/features/recommandation/components/bottomSheet/bottomSheet.stories.ts
@@ -1,5 +1,5 @@
-import recommendedSpotItemListMock from '../../mocks/recommendedSpotItemListMock';
-import startingSpotNameListMock from '../../mocks/startingSpotNameListMock';
+import recommendedSpotItemListMock from '../../../../mocks/recommendedSpotItemListMock';
+import startingSpotNameListMock from '../../../../mocks/startingSpotNameListMock';
 
 import BottomSheet from './BottomSheet';
