+++ frontend/src/features/recommandation/components/spotItemList/SpotItemList.tsx
