+++ frontend/src/features/recommandation/components/spotItemList/spotItemList.styled.ts
